+++ src/utility/StringUtility.h
@@ -18,7 +18,7 @@ namespace utility {
 	std::vector<std::string> wrap_text(const std::string &text, const int font_size, rendering::Renderer *renderer, const SDL_Rect &text_box_size, const std::pair<int, int> offsets_from_sides);
 
 	// TODO: replace with unicode library that handles this for us...
-	std::string substring_utf8(const std::string &text, const unsigned int  pos, const unsigned int length);
+	std::string substring_utf8(const std::string &text, const size_t pos, const size_t length);
 	std::string erase_from_end_utf8(const std::string &text, const int erasable_char_count);
 	// deletes n character from the position; so characters position - n -> position are erased
 	std::string erase_from_before_position_utf8(const std::string &text, const int erasable_char_count, const int position);
